move `Client` struct into `signaling` crate
