+++ client/Cargo.toml
@@ -7,18 +7,7 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0.95"
-reqwest = { version = "0.12.9", features = ["blocking", "json"] }
 signaling = { path = "../signaling" }
-str0m = "0.6.3"
 tokio = { version = "1.41.1", features = ["full"] }
 tracing = "0.1.41"
 tracing-subscriber = "0.3.18"
-
-[dependencies.uuid]
-version = "1.11.0"
-features = [
-    "v4",                # Lets you generate random UUIDs
-    "fast-rng",          # Use a faster (but still sufficiently random) RNG
-    "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
-    "serde",
-]
