+++ signaling/Cargo.toml
@@ -6,13 +6,16 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = "1.0.95"
 once_cell = "1.20.2"
+reqwest = { version = "0.12.9", features = ["blocking", "json"] }
 rouille = { version = "3.6.2", features = ["ssl"] }
 rustls = "0.23.19"
 serde = "1.0.217"
 serde_json = "1.0.133"
 str0m = "0.6.3"
 systemstat = "0.2.3"
+tokio = { version = "1.41.1", features = ["full"] }
 tracing = "0.1.41"
 tracing-subscriber = "0.3.18"
