+++ signaling/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod client;
 pub mod message;
 pub mod util;
